refactor(metadata): extract Hardcover list handling
